Move solvers/qpSpline to common/trajectories/qp_spline and lowercase its files
