+++ drake/common/trajectories/qp_spline/continuity_constraint.h
