+++ drake/common/trajectories/qp_spline/test/spline_generation_test.cc
@@ -1,4 +1,4 @@
-#include "drake/solvers/qpSpline/splineGeneration.h"
+#include "drake/common/trajectories/qp_spline/spline_generation.h"
 
 #include <vector>
