+++ drake/common/trajectories/qp_spline/value_constraint.cc
@@ -1,4 +1,4 @@
-#include "drake/solvers/qpSpline/ValueConstraint.h"
+#include "drake/common/trajectories/qp_spline/value_constraint.h"
 
 #include "drake/common/drake_assert.h"
