+++ drake/common/trajectories/qp_spline/value_constraint.h
