+++ drake/matlab/solvers/qpSpline/nWaypointCubicSplinemex.cpp
@@ -1,9 +1,11 @@
 #include <mex.h>
 
-#include <Eigen/Core>
-#include "drake/solvers/qpSpline/splineGeneration.h"
-#include "drake/matlab/util/drakeMexUtil.h"
 #include <iostream>
+
+#include <Eigen/Core>
+
+#include "drake/common/trajectories/qp_spline/spline_generation.h"
+#include "drake/matlab/util/drakeMexUtil.h"
 
 using namespace std;
 using namespace Eigen;
