--- drake/solvers/qpSpline/test/CMakeLists.txt
@@ -1,2 +1,0 @@
-drake_add_cc_test(NAME testSplineGeneration EXTENSION cpp)
-target_link_libraries(testSplineGeneration drakeSplineGeneration)
